requirements.txt
keyboard
numpy
pyautogui
opencv-python
Pillow
